requirements.txt
fastapi
uvicorn
groq
python-dotenv
firebase-admin